Add plucking rod experiment plots
In these plots we can see the underdamped response of the rod
vibrating in elevation, and in azimuth, with the ball fixed.

The frequencies are slightly lower than the artifact we were
seeing while flying.  We think this is caused by the extra ball
adding tension to the rod.

This phenomenon is minor in amplitude, but may need to be filtered
out in control experiments if they are sensitive to derivatives.